requirements/local.txt
-r testing.txt
django-debug-toolbar
coverage
django-coverage
flake8

